// projects/markets-web-app/src/generated/points-markets-api.ts
export interface components {
  schemas: {
// ...
    PublicPointPackageRevisionData: {
      pointPackageId: components["schemas"]["OpaqueId"];
      pointPackageRevisionId: components["schemas"]["OpaqueId"];
      /** @enum {string} */
      status: "ACTIVE" | "INACTIVE";
      name: string;
      description: string | null;
      /** Format: uri */
      relatedUrl: string | null;
      totalWeight: components["schemas"]["PositiveSafeInteger"];
      packageTick: components["schemas"]["PositiveSafeInteger"];
      contentHash: components["schemas"]["Sha256Hash"];
      components: components["schemas"]["PublicPointPackageRevisionComponent"][];
    };
// ...
    AuctionEligibilityRequest: {
      auctionCommandId: components["schemas"]["OpaqueId"];
      auctionCommandHash: components["schemas"]["Sha256Hash"];
      items: components["schemas"]["AuctionEligibilityRequestItem"][];
    };
// ...
    AuctionEligibilityResponse: {
      data: components["schemas"]["AuctionEligibilityData"];
      meta: components["schemas"]["RequestMeta"];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}